juscms/requirements.txt
Django==1.9.1
Pillow==3.1.0
django-mptt==0.8.0
